requirements.txt
alembic==1.16.1
annotated-types==0.7.0
dnspython==2.7.0
email_validator==2.2.0
greenlet==3.2.3
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
psycopg2==2.9.10
pydantic==2.11.7
pydantic-settings==2.9.1
pydantic_core==2.33.2
python-dotenv==1.1.0
SQLAlchemy==2.0.41
sqlmodel==0.0.24
typing-inspection==0.4.1
typing_extensions==4.14.0
wheel==0.45.1
